third_party/blink/web_tests/external/wpt/paint-timing/with-first-paint/sibling-painting-first-image-expected.txt

This is a testharness.js-based test.
FAIL Frame ignores paint-timing events fired from sibling frame. assert_equals: expected 1 but got 2
Harness: the test ran to completion.
